Highest prime factor

+++ 0x04-more_functions_nested_loops/100-prime_factor.c
@@ -7,7 +7,7 @@
 int main(void)
 {
 	int n, l;
-	long long int num = 612852475143;
+	long int num = 612852475143;
 
 	for (i = 2; i < num; i++)
 	{
